remove router functionality
