--- src/components/TheHeader.vue
@@ -1,21 +1,0 @@
-<template>
-  <header class="c-page-header">
-    <h1>Code Challenge</h1>
-  </header>
-</template>
-
-<script>
-export default {
-
-};
-</script>
-
-<style lang="scss">
-.c-page-header {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 97px;
-  border-bottom: solid 3px #3a3737;
-}
-</style>
