+++ src/router.js
@@ -1,8 +1,8 @@
-import Vue from 'vue';
+/* import Vue from 'vue';
 import Router from 'vue-router';
-import Home from './views/Home.vue';
 
-Vue.use(Router);
+
+// Vue.use(Router);
 
 export default new Router({
   mode: 'history',
@@ -15,3 +15,4 @@ export default new Router({
     },
   ],
 });
+*/
